requirements.txt
google-generativeai>=0.3.0
pydantic>=2.0.0
python-dateutil>=2.8.0
requests>=2.31.0

# API Server
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.0.0
